requirements.txt
discord.py[voice]
pytube
youtube-search-python
termcolor
pytz